requirements.txt
flask
pyyaml
requests
flask-socketio
pathlib
asyncio
websocket-client